server/requirements.txt
# pymongo[snappy,gssapi,srv,tls]
pymongo
dnspython
djongo
djangorestframework
requests
Django
Pillow
gunicorn
python-dotenv
